Made a LocalDataGenerator that is used in the E2ETestController and LocalConfig
